Rename dependency manager
